Add content to your recipe page for sloppy joes
I added an image of a sloppy joe. Then I added headings for each section
after the image. I added a short description of a sloppy joe. I then
used an unordered list to create the ingredients section content and an
ordered list to create the steps for making a sloppy joe.