Add CSS to the Odin Recipes project from The Odin project
I added a css directory and then created a recipes.css file to add
styling to the recipes pages. I then linked the css file to all of the
recipe html files. By using the same css file this allowed me to make
changes and keep the formatting for all of my recipe pages. This was
just a way to play around with CSS and try some of the skills I'm
learning.